ldn/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pycparser==2.21
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
